Close the confused-deputy escalation, ambiguous RBAC target, payroll race; basic_pay -> decimal(15,2)
Four findings, three of them from an independent review of this PR that the
round 1-8 review did not contain.

F-001 (High, security) — confused-deputy privilege escalation.
syncSpatieRolesFromWizardSelection()'s Super Admin gate was correct, but nothing
protected the data it trusts. member/update carried only `auth`, the controller
declared no middleware or policy, every StoreMemberStep*Request::authorize()
returns true, and emp_id arrived in the request body unvalidated. A zero-role
account could rewrite any employee's record and seed employee_role_mapping for
them — the table the edit form pre-checks Step 3 from, so the next privileged
save turned the seeded row into a live Spatie role. Reproduced end to end before
the fix (HTTP 200, victim renamed, role granted); the same probe now returns 403
with the victim record untouched and no mapping row written.

  - authorizeMemberWrite(): Super Admin may write any member; anyone else only
    their own employee record. Chosen after checking the data: every
    member/employee permission row is held by Super Admin and no other role, so
    this matches the access model the permission table already describes.
  - emp_id validated (required|integer|exists) before use, so an unknown id is a
    422 rather than a fatal on find()->update().
  - store(), toggleStatus(), destroy() are admin-only — no self-service case.
  - the employee_role_mapping write is gated on actingUserCanManageRbacRoles(),
    which is what actually breaks the seeding half of the chain.

  Self-service profile editing is deliberately NOT locked out: the header's Edit
  Profile link posts to this same endpoint. That path separately answers 422
  because the profile form sends no userrole[] while combinedMemberRules() still
  marks it required — a pre-existing bug, left alone and flagged, not masked.

F-002 (High, security) — user_credentials.user_id is not unique (611 duplicate
groups on the live data, 207 of them real employees, 603 of the affected rows
already holding Spatie roles), so where('user_id', ...)->first() handed
syncRoles() an arbitrary one of several logins. The sync now refuses and logs for
an ambiguous member rather than granting or revoking on a login nobody chose, and
the lookup is ordered so it is at least stable. Reconciling the duplicates
remains a DBA task.

F-006 (Medium, concurrency) — saveStep6PayrollData() was a read-then-write with
no unique key behind it; under REPEATABLE READ a plain SELECT takes no lock, so
two concurrent saves could both insert. Adds a unique index on
payroll_salary_master.employee_master_pk (migration refuses, naming the rows, if
duplicates already exist) and switches to updateOrCreate.

F-009 (Medium, data integrity) — basic_pay was the one money column on
payroll_salary_master that could not hold paise; every sibling is double(15,2).
Decision taken: basic pay may carry paise. Column is decimal(15,2) — exact rather
than binary floating point, which is the right choice for money. Original
migration corrected for fresh installs, new migration converts environments where
it already ran, rule numeric instead of integer, and both step-6 inputs carry
step="0.01".

Tests: 210 total (+14), 1257 assertions. The only failures are the 3 pre-existing
FcStepReportGuardsTest cases, which fail identically at 014e0324d.

+++ database/migrations/2026_08_19_000001_add_basic_pay_to_payroll_salary_master.php
@@ -12,7 +12,13 @@ return new class extends Migration
         // by the Estate module — this only adds the one column the new Member wizard step needs.
         if (! Schema::hasColumn('payroll_salary_master', 'basic_pay')) {
             Schema::table('payroll_salary_master', function (Blueprint $table) {
-                $table->integer('basic_pay')->nullable()->after('employee_category_master_pk');
+                // decimal(15,2), matching every other money column on this table
+                // (net_salary, gross_salary, tds, payable_amount, deduction_amount,
+                // tds_adjust). A plain integer() here could not represent paise and was
+                // the only money column on the table that could not — PR #319 review,
+                // F-009. Environments where this migration already ran with the narrower
+                // type are corrected by 2026_09_22_000001.
+                $table->decimal('basic_pay', 15, 2)->nullable()->after('employee_category_master_pk');
             });
         }
     }
